[opt] physics2d box2d contactListner optimization
